change mcp concepts from utils to their own directories
