+++ src/api/resources/__init__.py
@@ -1,0 +1,4 @@
+from .resources import resources
+from .register import register_resources
+
+__all__ = ["resources", "register_resources"]
