+++ src/api/tools/__init__.py
@@ -1,0 +1,4 @@
+from .tools import tools
+from .register import register_tools
+
+__all__ = ["tools", "register_tools"]
